added completed transactions and changed prices from int to decimal

+++ lib/model/app/future_payment.dart
@@ -1,9 +1,10 @@
+import 'package:decimal/decimal.dart';
 import 'package:trippidy/model/member.dart';
 
 class FuturePayment {
   final Member payer;
   final Member payee;
-  final double amount;
+  final Decimal amount;
 
   FuturePayment(this.payer, this.payee, this.amount);
 }
